[ROS] Remove hector_geotiff from build to remove qt4 dep
