+++ controller_wr_driver/src/link_def.c
@@ -1,7 +1,7 @@
 #include <common.h>
 
 #include "mproto.h"
-#include "mproto_def.h"
+#include "link_def.h"
 #include "uc_pc_link_defs.h"
 
 static const mproto_driver_cb_ctx_t default_cb_ctx = {
@@ -10,9 +10,6 @@ static const mproto_driver_cb_ctx_t default_cb_ctx = {
 
     .set_odom_params_cb     = NULL,
     .reset_odometry_cb      = NULL
-
-    // .get_control_params     = NULL,
-    // .set_control_params_cb  = NULL
 };
 
 static mproto_driver_cb_ctx_t last_cb_ctx;
@@ -22,71 +19,17 @@ mproto_driver_cb_ctx_t mproto_driver_get_new_cb_ctx( void )
     return default_cb_ctx;
 }
 
-// void trigger_task_cb( const std_srvs::TriggerRequest &req, std_srvs::TriggerResponse &resp )
-// {   
-//     (void)req;
-//     (void)resp;
-
-//     dbgprintf( "Called [%s]\n\r", __FUNCTION__ );
-
-//     resp.success = true;
-// }
-
-// void steer_params_cb( const wr8_msgs::SteerParamsRequest &req, wr8_msgs::SteerParamsResponse &resp )
-// {
-//     (void)req;
-//     (void)resp;
-
-//     dbgprintf( "Called [%s]\n\r", __FUNCTION__ );
-
-//     if ( last_cb_ctx.set_odom_params_cb )
-//     {
-//         last_cb_ctx.set_odom_params_cb( req.left_k, req.right_k );
-//     }
-// }
-
-// void control_params_cb( const wr8_msgs::ControlParamsRequest &req, wr8_msgs::ControlParamsResponse &resp )
-// {
-//     (void)req;
-//     (void)resp;
-
-//     dbgprintf( "Called %d [%s]\n\r", req.request_only, __FUNCTION__ );
-
-//     // if ( req.request_only && last_cb_ctx.get_control_params )
-//     // {
-//     //     control_params_setup_t params = last_cb_ctx.get_control_params();
-
-//     //     static const size_t data_length = 2;
-//     //     static float data[data_length];
-
-//     //     data[0] = params.esc_min_dc_offset;
-//     //     data[1] = params.esc_max_dc_offset;
-
-//     //     resp.params = data;
-//     //     resp.params_length = data_length;
-//     // }
-
-//     // if ( last_cb_ctx.set_control_params_cb )
-//     // {
-//     //     control_params_setup_t params;
-//     //     params.esc_min_dc_offset = req.esc_min_dc_offset;
-//     //     params.esc_max_dc_offset = req.esc_max_dc_offset;
-
-//     //     last_cb_ctx.set_control_params_cb( &params );
-//     // }
-// }
-
 void reset_odometry_cb( void )
 {
-    if ( last_cb_ctx.reset_odometry_cb )
-    {
-        last_cb_ctx.reset_odometry_cb();
-    }
+    if ( !last_cb_ctx.reset_odometry_cb )
+        return;
+
+    last_cb_ctx.reset_odometry_cb();
 }
 
 void cmd_vel_cb( float data[2] )
 {
-    if ( last_cb_ctx.cmd_cb == NULL )
+    if ( !last_cb_ctx.cmd_cb )
         return;
 
     float *vel_data = (float *)data;
@@ -102,7 +45,7 @@ void cmd_vel_cb( float data[2] )
 
 void raw_cmd_vel_cb( float data[2] )
 {
-    if ( last_cb_ctx.cmd_cb == NULL )
+    if ( !last_cb_ctx.raw_cmd_cb )
         return;
 
     float *vel_data = (float *)data;
@@ -132,22 +75,6 @@ void cmd_cb( mpcmd_t cmd, uint8_t *data, size_t len )
         raw_cmd_vel_cb( data );
     }
 }
-
-// void raw_vel_cb( const geometry_msgs::Twist &msg )
-// {
-//     if ( last_cb_ctx.raw_cmd_cb == NULL )
-//     {
-//         return;
-//     }
-
-//     float cmd_speed = msg.linear.x;
-//     float cmd_steer = msg.angular.z;
-
-//     cmd_speed = CLIP_VALUE(cmd_speed, -100, 100);
-//     cmd_steer = CLIP_VALUE(cmd_steer, -100, 100);
-
-//     last_cb_ctx.raw_cmd_cb( cmd_speed, cmd_steer );
-// }
 
 /*
  * Spin thread - used to receive messages
@@ -202,7 +129,7 @@ void mproto_driver_send_pose( float x, float y, float dir, float vx, float uz )
     mproto_send_data(mproto_ctx, WR_OUT_CMD_ODOM_DATA, (uint8_t *)data, sizeof(data));
 }
 
-void mproto_driver_send_encoder_raw( int32_t value )
+void mproto_driver_send_encoder_raw( float value )
 {
     mproto_send_data(mproto_ctx, WR_OUT_CMD_ENCODER_VALUE, (uint8_t *)&value, sizeof(value));
 }
@@ -283,6 +210,7 @@ void mproto_driver_init( tprio_t prio, mproto_driver_cb_ctx_t *ctx )
 
     mproto_register_command( mproto_ctx, WR_IN_CMD_VEL, cmd_cb );
     mproto_register_command( mproto_ctx, WR_IN_CMD_RESET_ODOM, cmd_cb );
+    mproto_register_command( mproto_ctx, WR_IN_CMD_RAW_VEL, cmd_cb );
 
     chThdCreateStatic( waSpinner, sizeof(waSpinner), prio, Spinner, NULL );
 }
