+++ install_pkgs.sh
@@ -27,7 +27,6 @@ sudo apt install ros-$ROS_DISTRO-base-local-planner \
 					ros-$ROS_DISTRO-gmapping \
 					ros-$ROS_DISTRO-camera-info-manager \
 					ros-$ROS_DISTRO-roslint \
-					qt4-default \
 					libsuitesparse-dev
